initial stab at multi-chat
