+++ App_Function_Libraries/Gradio_Related.py
@@ -16,6 +16,8 @@ import gradio as gr
 # Local Imports
 from App_Function_Libraries.DB.DB_Manager import get_db_config
 from App_Function_Libraries.Gradio_UI.Audio_ingestion_tab import create_audio_processing_tab
+from App_Function_Libraries.Gradio_UI.Character_Interaction_tab import create_character_card_interaction_tab, \
+    create_multiple_character_chat_tab
 from App_Function_Libraries.Gradio_UI.Chat_ui import create_chat_management_tab, \
     create_chat_interface_four, create_chat_interface_multi_api, create_chat_interface_stacked, create_chat_interface
 from App_Function_Libraries.Gradio_UI.Config_tab import create_config_editor_tab
@@ -273,8 +275,9 @@ def launch_ui(share_public=None, server_mode=False):
                 create_chat_with_llamafile_tab()
                 create_chat_management_tab()
                 chat_workflows_tab()
-                from App_Function_Libraries.Gradio_UI.Writing_tab import create_character_card_interaction_tab
+                create_multiple_character_chat_tab()
                 create_character_card_interaction_tab()
+
 
             with gr.TabItem("View DB Items"):
                 create_viewing_tab()
